Release version 1.0.0
This release sets up a build task to compile the JSX and any ES6 syntax
into vanilla ES5 without JSX. The built version of the component is
saved to 'build/ReactWaypoint.js'.

To run the build task:
> npm run build

This release also contains a fix for a bug where the 'resize' event was
possibly missed when we had a 'scrollableParent' other than the window.